Bug fixed, needs a little bit of added work
Finished, it just needs to be "assimilated" into the main game code, for example it needs the ability to add a certain item to the player's inventory and it needs the main method removed so that it'll work with the rest of the code. The main method was only there for testing, and should prove no problem to remove since all that's in main is the method call.